Updated rep-tile

+++ Backend/smart_wrappers/general_utils.cpp
@@ -231,8 +231,8 @@ short Dtest_equality(double* C_comp, double* C, long long size) {
   return (short) acc;
 }
 
-inline double Herror(__half a, __half b) {
-  if (a == 0) return (__half) dabs(a - b);
+inline __half Herror(__half a, __half b) {
+  if (a == ((__half) 0)) return (__half) dabs(a - b);
   return (__half) dabs(a - b)/a;
 }
 
@@ -249,9 +249,9 @@ long int Hvec_diff(__half* a, __half* b, long long size, __half eps) {
 
 short Htest_equality(__half* C_comp, __half* C, long long size) {
   long int acc = 2, failed;
-  float eps = 1e-2;
+  __half eps = 1e-2;
   failed = Hvec_diff(C_comp, C, size, eps);
-  while (eps > FLT_MIN && !failed && acc < 30) {
+  while (eps > (__half) FLT_MIN && !failed && acc < 30) {
     eps *= 0.1;
     acc++;
     failed = Hvec_diff(C_comp, C, size, eps);
